+++ HISTORY.rst
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+1.3.0-dev
+---------
+
+**API Changes**
+
+- Throw ``BadWeightError`` when creating or reprioritising a stream with a
+  weight that is not an integer between 1 and 256, inclusive.
+
 1.2.2 (2016-11-11)
 ------------------